Add Mojave assets and Dark Mode support
This changeset grabs the Mojave assets from upstream and uses them on
macOS versions later than 10.14, applying a check for Dark Mode on said
assets. It also updates to use the upstream Yosemite assets at the same
display sizes (26px height for the bezel; 8x8 and 10x10 for the cancel
and snapback assets). Since there are no same-sized assets for Snow
Leopard upstream, support for Snow Leopard-styled assets has been
removed. (Jumpcut 0.7x doesn't support pre-Yosemite macOS,
regardless.)

This is probably the last go-round for this fork of the ShortcutRecord2
code; in the event future substantive changes are required, strongly
consider making the changes required to support mainline
ShortcutRecorder2 or switching to a different hotkey library, such as
MASShortcut.